added the 5th and 6th svgs
