+++ src/components/why-choose-us.js
@@ -1,369 +1,232 @@
 import React from "react";
 
-export default function WhyChooseUs() {
+function twoByTwo() {
   return (
-    <section className="flex justify-center mx-auto max-w-screen-xl py-8">
-      <ul className="grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-3">
-        <li className="col-span-1 bg-white rounded-lg shadow">
-          <div className="w-full flex items-center justify-between p-6 space-x-6">
-            <div className="flex-1 truncate">
-              <div className="flex items-center space-x-3">
-                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
-                  Jane Cooper
-                </h3>
-                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
-                  Admin
-                </span>
-              </div>
-              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
-                Regional Paradigm Technician
-              </p>
-            </div>
-            <img
-              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
-              src="https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
-              alt=""
-            />
-          </div>
-          <div className="border-t border-gray-200">
-            <div className="-mt-px flex">
-              <div className="w-0 flex-1 flex border-r border-gray-200">
-                <a
-                  href="#"
-                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
-                >
-                  {/* // <!-- Heroicon name: mail --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
-                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
-                  </svg>
-                  <span className="ml-3">Email</span>
-                </a>
-              </div>
-              <div className="-ml-px w-0 flex-1 flex">
-                <a
-                  href="#"
-                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
-                >
-                  {/* // <!-- Heroicon name: phone --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
-                  </svg>
-                  <span className="ml-3">Call</span>
-                </a>
-              </div>
-            </div>
-          </div>
-        </li>
+    <div className="bg-gray-50 overflow-hidden">
+      <div className="relative max-w-screen-xl mx-auto py-12 px-6">
+        <svg
+          className="absolute top-0 left-full transform -translate-x-1/2 -translate-y-3/4 lg:left-auto lg:right-full lg:translate-x-2/3 lg:translate-y-1/4"
+          width="404"
+          height="784"
+          fill="none"
+          viewBox="0 0 404 784"
+        >
+          <defs>
+            <pattern
+              id="8b1b5f72-e944-4457-af67-0c6d15a99f38"
+              x="0"
+              y="0"
+              width="20"
+              height="20"
+              patternUnits="userSpaceOnUse"
+            >
+              <rect
+                x="0"
+                y="0"
+                width="4"
+                height="4"
+                className="text-gray-200"
+                fill="currentColor"
+              />
+            </pattern>
+          </defs>
+          <rect
+            width="404"
+            height="784"
+            fill="url(#8b1b5f72-e944-4457-af67-0c6d15a99f38)"
+          />
+        </svg>
 
-        <li className="col-span-1 bg-white rounded-lg shadow">
-          <div className="w-full flex items-center justify-between p-6 space-x-6">
-            <div className="flex-1 truncate">
-              <div className="flex items-center space-x-3">
-                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
-                  Cody Fisher
-                </h3>
-                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
-                  Admin
-                </span>
-              </div>
-              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
-                Product Directives Officer
-              </p>
-            </div>
-            <img
-              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
-              src="https://images.unsplash.com/photo-1570295999919-56ceb5ecca61?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
-              alt=""
-            />
+        <div className="relative lg:grid lg:grid-cols-3 lg:gap-x-8">
+          <div className="lg:col-span-1">
+            <h3 className="text-3xl leading-9 font-extrabold tracking-tight text-gray-900 sm:text-4xl sm:leading-10">
+              Why choose us?
+            </h3>
           </div>
-          <div className="border-t border-gray-200">
-            <div className="-mt-px flex">
-              <div className="w-0 flex-1 flex border-r border-gray-200">
-                <a
-                  href="#"
-                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+          <div className="mt-10 sm:grid sm:grid-cols-3 sm:gap-x-8 sm:gap-y-10 lg:col-span-2">
+            <div>
+              <div className="flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                {/* <!-- Heroicon name: globe-alt --> */}
+                <svg
+                  className="h-6 w-6"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  stroke="currentColor"
                 >
-                  {/* // <!-- Heroicon name: mail --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
-                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
-                  </svg>
-                  <span className="ml-3">Email</span>
-                </a>
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth="2"
+                    d="M21 12a9 9 0 01-9 9m9-9a9 9 0 00-9-9m9 9H3m9 9a9 9 0 01-9-9m9 9c1.657 0 3-4.03 3-9s-1.343-9-3-9m0 18c-1.657 0-3-4.03-3-9s1.343-9 3-9m-9 9a9 9 0 019-9"
+                  />
+                </svg>
               </div>
-              <div className="-ml-px w-0 flex-1 flex">
-                <a
-                  href="#"
-                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
-                >
-                  {/* // <!-- Heroicon name: phone --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
-                  </svg>
-                  <span className="ml-3">Call</span>
-                </a>
+              <div className="mt-5">
+                <h4 className="text-lg leading-6 font-medium text-gray-900">
+                  Specialized Company
+                </h4>
+                <p className="mt-2 text-base leading-6 text-gray-500">
+                  We are a landscaping company that specializes in residential
+                  and commercial landscaping
+                </p>
               </div>
             </div>
-          </div>
-        </li>
-
-        <li className="col-span-1 bg-white rounded-lg shadow">
-          <div className="w-full flex items-center justify-between p-6 space-x-6">
-            <div className="flex-1 truncate">
-              <div className="flex items-center space-x-3">
-                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
-                  Esther Howard
-                </h3>
-                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
-                  Admin
-                </span>
-              </div>
-              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
-                Forward Response Developer
-              </p>
-            </div>
-            <img
-              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
-              src="https://images.unsplash.com/photo-1520813792240-56fc4a3765a7?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
-              alt=""
-            />
-          </div>
-          <div className="border-t border-gray-200">
-            <div className="-mt-px flex">
-              <div className="w-0 flex-1 flex border-r border-gray-200">
-                <a
-                  href="#"
-                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+            <div className="mt-10 sm:mt-0">
+              <div className="flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                {/* <!-- Heroicon name: scale --> */}
+                <svg
+                  className="h-6 w-6"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  stroke="currentColor"
                 >
-                  {/* // <!-- Heroicon name: mail --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
-                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
-                  </svg>
-                  <span className="ml-3">Email</span>
-                </a>
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth="2"
+                    d="M3 6l3 1m0 0l-3 9a5.002 5.002 0 006.001 0M6 7l3 9M6 7l6-2m6 2l3-1m-3 1l-3 9a5.002 5.002 0 006.001 0M18 7l3 9m-3-9l-6-2m0-2v2m0 16V5m0 16H9m3 0h3"
+                  />
+                </svg>
               </div>
-              <div className="-ml-px w-0 flex-1 flex">
-                <a
-                  href="#"
-                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
-                >
-                  {/* // <!-- Heroicon name: phone --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
-                  </svg>
-                  <span className="ml-3">Call</span>
-                </a>
+              <div className="mt-5">
+                <h4 className="text-lg leading-6 font-medium text-gray-900">
+                  Licensed & Insured
+                </h4>
+                <p className="mt-2 text-base leading-6 text-gray-500">
+                  All our landscapers are fully licensed, bonded and insured for
+                  their safety
+                </p>
               </div>
             </div>
-          </div>
-        </li>
-
-        <li className="col-span-1 bg-white rounded-lg shadow">
-          <div className="w-full flex items-center justify-between p-6 space-x-6">
-            <div className="flex-1 truncate">
-              <div className="flex items-center space-x-3">
-                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
-                  Jenny Wilson
-                </h3>
-                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
-                  Admin
-                </span>
-              </div>
-              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
-                Central Security Manager
-              </p>
-            </div>
-            <img
-              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
-              src="https://images.unsplash.com/photo-1498551172505-8ee7ad69f235?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
-              alt=""
-            />
-          </div>
-          <div className="border-t border-gray-200">
-            <div className="-mt-px flex">
-              <div className="w-0 flex-1 flex border-r border-gray-200">
-                <a
-                  href="#"
-                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+            <div className="mt-10 sm:mt-0">
+              <div className="flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                {/* <!-- Heroicon name: lightning-bolt --> */}
+                <svg
+                  className="h-6 w-6"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  stroke="currentColor"
                 >
-                  {/* // <!-- Heroicon name: mail --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
-                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
-                  </svg>
-                  <span className="ml-3">Email</span>
-                </a>
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth="2"
+                    d="M13 10V3L4 14h7v7l9-11h-7z"
+                  />
+                </svg>
               </div>
-              <div className="-ml-px w-0 flex-1 flex">
-                <a
-                  href="#"
-                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
-                >
-                  {/* // <!-- Heroicon name: phone --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
-                  </svg>
-                  <span className="ml-3">Call</span>
-                </a>
+              <div className="mt-5">
+                <h4 className="text-lg leading-6 font-medium text-gray-900">
+                  Dependable Services
+                </h4>
+                <p className="mt-2 text-base leading-6 text-gray-500">
+                  We love to take pride in the work we do. Each project is
+                  finished in time and budget
+                </p>
               </div>
             </div>
-          </div>
-        </li>
-
-        <li className="col-span-1 bg-white rounded-lg shadow">
-          <div className="w-full flex items-center justify-between p-6 space-x-6">
-            <div className="flex-1 truncate">
-              <div className="flex items-center space-x-3">
-                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
-                  Kristin Watson
-                </h3>
-                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
-                  Admin
-                </span>
+            <div className="mt-10 sm:mt-0">
+              <div className="flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                {/* <!-- Heroicon name: mail --> */}
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  stroke="currentColor"
+                  className="h-6 w-6"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={2}
+                    d="M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z"
+                  />
+                </svg>
               </div>
-              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
-                Lead Implementation Liaison
-              </p>
+              <div className="mt-5">
+                <h4 className="text-lg leading-6 font-medium text-gray-900">
+                  Day Scheduling
+                </h4>
+                <p className="mt-2 text-base leading-6 text-gray-500">
+                  We schedule regular appointments to visit your property on the
+                  same day and time of the week
+                </p>
+              </div>
             </div>
-            <img
-              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
-              src="https://images.unsplash.com/photo-1532417344469-368f9ae6d187?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
-              alt=""
-            />
-          </div>
-          <div className="border-t border-gray-200">
-            <div className="-mt-px flex">
-              <div className="w-0 flex-1 flex border-r border-gray-200">
-                <a
-                  href="#"
-                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
-                >
-                  {/* // <!-- Heroicon name: mail --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
+            <div className="mt-10 sm:mt-0">
+              <div className="flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                {/* <!-- Heroicon name: mail --> */}
+                {/* <svg
+                    className="h-6 w-6"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    stroke="currentColor"
                   >
-                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
-                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
-                  </svg>
-                  <span className="ml-3">Email</span>
-                </a>
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth="2"
+                      d="M3 8l7.89 5.26a2 2 0 002.22 0L21 8M5 19h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z"
+                    />
+                  </svg> */}
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  stroke="currentColor"
+                  className="h-6 w-6"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={2}
+                    d="M7 8h10M7 12h4m1 8l-4-4H5a2 2 0 01-2-2V6a2 2 0 012-2h14a2 2 0 012 2v8a2 2 0 01-2 2h-3l-4 4z"
+                  />
+                </svg>
               </div>
-              <div className="-ml-px w-0 flex-1 flex">
-                <a
-                  href="#"
-                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+              <div className="mt-5">
+                <h4 className="text-lg leading-6 font-medium text-gray-900">
+                  Free Consultations
+                </h4>
+                <p className="mt-2 text-base leading-6 text-gray-500">
+                  We offer free consultations for our services, and will provide
+                  you with an actual quote
+                </p>
+              </div>
+            </div>
+            <div className="mt-10 sm:mt-0">
+              <div className="flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                {/* <!-- Heroicon name: mail --> */}
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  stroke="currentColor"
+                  className="w-6 h-6"
                 >
-                  {/* // <!-- Heroicon name: phone --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
-                  </svg>
-                  <span className="ml-3">Call</span>
-                </a>
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={2}
+                    d="M9 5H7a2 2 0 00-2 2v12a2 2 0 002 2h10a2 2 0 002-2V7a2 2 0 00-2-2h-2M9 5a2 2 0 002 2h2a2 2 0 002-2M9 5a2 2 0 012-2h2a2 2 0 012 2m-6 9l2 2 4-4"
+                  />
+                </svg>
+              </div>
+              <div className="mt-5">
+                <h4 className="text-lg leading-6 font-medium text-gray-900">
+                  Reputable Company
+                </h4>
+                <p className="mt-2 text-base leading-6 text-gray-500">
+                  Operating for more than 30 years, earning a repuation for
+                  service and beautiful work
+                </p>
               </div>
             </div>
           </div>
-        </li>
-
-        <li className="col-span-1 bg-white rounded-lg shadow">
-          <div className="w-full flex items-center justify-between p-6 space-x-6">
-            <div className="flex-1 truncate">
-              <div className="flex items-center space-x-3">
-                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
-                  Cameron Williamson
-                </h3>
-                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
-                  Admin
-                </span>
-              </div>
-              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
-                Internal Applications Engineer
-              </p>
-            </div>
-            <img
-              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
-              src="https://images.unsplash.com/photo-1566492031773-4f4e44671857?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
-              alt=""
-            />
-          </div>
-          <div className="border-t border-gray-200">
-            <div className="-mt-px flex">
-              <div className="w-0 flex-1 flex border-r border-gray-200">
-                <a
-                  href="#"
-                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
-                >
-                  {/* // <!-- Heroicon name: mail --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
-                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
-                  </svg>
-                  <span className="ml-3">Email</span>
-                </a>
-              </div>
-              <div className="-ml-px w-0 flex-1 flex">
-                <a
-                  href="#"
-                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
-                >
-                  {/* // <!-- Heroicon name: phone --> */}
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                  >
-                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
-                  </svg>
-                  <span className="ml-3">Call</span>
-                </a>
-              </div>
-            </div>
-          </div>
-        </li>
-      </ul>
-    </section>
+        </div>
+      </div>
+    </div>
   );
 }
+
+export default twoByTwo;
