+++ src/components/z-why-choose-us.js
@@ -1,0 +1,369 @@
+import React from "react";
+
+export default function WhyChooseUs() {
+  return (
+    <section className="flex justify-center mx-auto max-w-screen-xl py-8">
+      <ul className="grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-3">
+        <li className="col-span-1 bg-white rounded-lg shadow">
+          <div className="w-full flex items-center justify-between p-6 space-x-6">
+            <div className="flex-1 truncate">
+              <div className="flex items-center space-x-3">
+                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
+                  Jane Cooper
+                </h3>
+                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
+                  Admin
+                </span>
+              </div>
+              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
+                Regional Paradigm Technician
+              </p>
+            </div>
+            <img
+              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
+              src="https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
+              alt=""
+            />
+          </div>
+          <div className="border-t border-gray-200">
+            <div className="-mt-px flex">
+              <div className="w-0 flex-1 flex border-r border-gray-200">
+                <a
+                  href="#"
+                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: mail --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
+                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
+                  </svg>
+                  <span className="ml-3">Email</span>
+                </a>
+              </div>
+              <div className="-ml-px w-0 flex-1 flex">
+                <a
+                  href="#"
+                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: phone --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
+                  </svg>
+                  <span className="ml-3">Call</span>
+                </a>
+              </div>
+            </div>
+          </div>
+        </li>
+
+        <li className="col-span-1 bg-white rounded-lg shadow">
+          <div className="w-full flex items-center justify-between p-6 space-x-6">
+            <div className="flex-1 truncate">
+              <div className="flex items-center space-x-3">
+                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
+                  Cody Fisher
+                </h3>
+                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
+                  Admin
+                </span>
+              </div>
+              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
+                Product Directives Officer
+              </p>
+            </div>
+            <img
+              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
+              src="https://images.unsplash.com/photo-1570295999919-56ceb5ecca61?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
+              alt=""
+            />
+          </div>
+          <div className="border-t border-gray-200">
+            <div className="-mt-px flex">
+              <div className="w-0 flex-1 flex border-r border-gray-200">
+                <a
+                  href="#"
+                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: mail --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
+                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
+                  </svg>
+                  <span className="ml-3">Email</span>
+                </a>
+              </div>
+              <div className="-ml-px w-0 flex-1 flex">
+                <a
+                  href="#"
+                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: phone --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
+                  </svg>
+                  <span className="ml-3">Call</span>
+                </a>
+              </div>
+            </div>
+          </div>
+        </li>
+
+        <li className="col-span-1 bg-white rounded-lg shadow">
+          <div className="w-full flex items-center justify-between p-6 space-x-6">
+            <div className="flex-1 truncate">
+              <div className="flex items-center space-x-3">
+                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
+                  Esther Howard
+                </h3>
+                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
+                  Admin
+                </span>
+              </div>
+              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
+                Forward Response Developer
+              </p>
+            </div>
+            <img
+              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
+              src="https://images.unsplash.com/photo-1520813792240-56fc4a3765a7?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
+              alt=""
+            />
+          </div>
+          <div className="border-t border-gray-200">
+            <div className="-mt-px flex">
+              <div className="w-0 flex-1 flex border-r border-gray-200">
+                <a
+                  href="#"
+                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: mail --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
+                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
+                  </svg>
+                  <span className="ml-3">Email</span>
+                </a>
+              </div>
+              <div className="-ml-px w-0 flex-1 flex">
+                <a
+                  href="#"
+                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: phone --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
+                  </svg>
+                  <span className="ml-3">Call</span>
+                </a>
+              </div>
+            </div>
+          </div>
+        </li>
+
+        <li className="col-span-1 bg-white rounded-lg shadow">
+          <div className="w-full flex items-center justify-between p-6 space-x-6">
+            <div className="flex-1 truncate">
+              <div className="flex items-center space-x-3">
+                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
+                  Jenny Wilson
+                </h3>
+                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
+                  Admin
+                </span>
+              </div>
+              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
+                Central Security Manager
+              </p>
+            </div>
+            <img
+              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
+              src="https://images.unsplash.com/photo-1498551172505-8ee7ad69f235?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
+              alt=""
+            />
+          </div>
+          <div className="border-t border-gray-200">
+            <div className="-mt-px flex">
+              <div className="w-0 flex-1 flex border-r border-gray-200">
+                <a
+                  href="#"
+                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: mail --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
+                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
+                  </svg>
+                  <span className="ml-3">Email</span>
+                </a>
+              </div>
+              <div className="-ml-px w-0 flex-1 flex">
+                <a
+                  href="#"
+                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: phone --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
+                  </svg>
+                  <span className="ml-3">Call</span>
+                </a>
+              </div>
+            </div>
+          </div>
+        </li>
+
+        <li className="col-span-1 bg-white rounded-lg shadow">
+          <div className="w-full flex items-center justify-between p-6 space-x-6">
+            <div className="flex-1 truncate">
+              <div className="flex items-center space-x-3">
+                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
+                  Kristin Watson
+                </h3>
+                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
+                  Admin
+                </span>
+              </div>
+              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
+                Lead Implementation Liaison
+              </p>
+            </div>
+            <img
+              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
+              src="https://images.unsplash.com/photo-1532417344469-368f9ae6d187?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
+              alt=""
+            />
+          </div>
+          <div className="border-t border-gray-200">
+            <div className="-mt-px flex">
+              <div className="w-0 flex-1 flex border-r border-gray-200">
+                <a
+                  href="#"
+                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: mail --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
+                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
+                  </svg>
+                  <span className="ml-3">Email</span>
+                </a>
+              </div>
+              <div className="-ml-px w-0 flex-1 flex">
+                <a
+                  href="#"
+                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: phone --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
+                  </svg>
+                  <span className="ml-3">Call</span>
+                </a>
+              </div>
+            </div>
+          </div>
+        </li>
+
+        <li className="col-span-1 bg-white rounded-lg shadow">
+          <div className="w-full flex items-center justify-between p-6 space-x-6">
+            <div className="flex-1 truncate">
+              <div className="flex items-center space-x-3">
+                <h3 className="text-gray-900 text-sm leading-5 font-medium truncate">
+                  Cameron Williamson
+                </h3>
+                <span className="flex-shrink-0 inline-block px-2 py-0.5 text-teal-800 text-xs leading-4 font-medium bg-teal-100 rounded-full">
+                  Admin
+                </span>
+              </div>
+              <p className="mt-1 text-gray-500 text-sm leading-5 truncate">
+                Internal Applications Engineer
+              </p>
+            </div>
+            <img
+              className="w-10 h-10 bg-gray-300 rounded-full flex-shrink-0"
+              src="https://images.unsplash.com/photo-1566492031773-4f4e44671857?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=4&amp;w=256&amp;h=256&amp;q=60"
+              alt=""
+            />
+          </div>
+          <div className="border-t border-gray-200">
+            <div className="-mt-px flex">
+              <div className="w-0 flex-1 flex border-r border-gray-200">
+                <a
+                  href="#"
+                  className="relative -mr-px w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-bl-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: mail --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
+                    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
+                  </svg>
+                  <span className="ml-3">Email</span>
+                </a>
+              </div>
+              <div className="-ml-px w-0 flex-1 flex">
+                <a
+                  href="#"
+                  className="relative w-0 flex-1 inline-flex items-center justify-center py-4 text-sm leading-5 text-gray-700 font-medium border border-transparent rounded-br-lg hover:text-gray-500 focus:outline-none focus:shadow-outline-blue focus:border-blue-300 focus:z-10 transition ease-in-out duration-150"
+                >
+                  {/* // <!-- Heroicon name: phone --> */}
+                  <svg
+                    className="w-5 h-5 text-gray-400"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                  >
+                    <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z" />
+                  </svg>
+                  <span className="ml-3">Call</span>
+                </a>
+              </div>
+            </div>
+          </div>
+        </li>
+      </ul>
+    </section>
+  );
+}
