+++ src/pages/index.js
@@ -2,7 +2,6 @@ import React from "react";
 
 import Layout from "../components/layout";
 import SEO from "../components/seo";
-import TwoByTwo from "../components/two-by-two";
 // import BetterCarousel from "../components/better-carousel";
 import AboutUs from "../components/about-us";
 import Carousel from "../components/carousel";
@@ -18,9 +17,9 @@ function IndexPage() {
       />
       <Carousel />
       <AboutUs />
-      <TwoByTwo />
-      <ExceptionalService />
       <WhyChooseUs/>
+      <ExceptionalService />
+      
     </Layout>
   );
 }
